SimpleController: fix sending commands without input samples
If a data connection to the input port exists the controller sends old
data as new commands even if no new input samples are available. This
results in a dangerous behaviour commanding the robot to move without
any timeout.

The new implementation repeats the command until a timeout
interval passed. After this stop commands are sent until a new
input sample arrived. This ensures that if the commanding component
freezes or forgets to send a stop command the robot stops after n
seconds.